backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
reportlab==4.0.9
PyPDF2==3.0.1
pillow==10.2.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
openai==1.12.0
anthropic==0.18.1
alembic==1.18.1
sqlalchemy==2.0.45
requests==2.31.0
python-docx==1.1.0
matplotlib==3.8.2
google-generativeai==0.3.2
